Move debugging code of index query to a separate module
